% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention.R
\name{attention}
\alias{attention}
\title{Self-Attention Algorithm}
\description{
Implements functions for the Self-Attention algorithm, and demonstrates with vignettes of increasing depth how to construct the Self-Attention algorithm..
}
\references{
{https://qua.st/attention/}
}
\seealso{
https://qua.st/attention
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
}
