% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeWeights.R
\name{ComputeWeights}
\alias{ComputeWeights}
\title{SoftMax sigmoid function}
\usage{
ComputeWeights(scores, K)
}
\arguments{
\item{scores}{input value (numeric)}

\item{K}{Key values matrix (numeric)}
}
\value{
output value (numeric)
}
\description{
SoftMax sigmoid function
}
