% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftMax.R
\name{SoftMax}
\alias{SoftMax}
\title{SoftMax sigmoid function}
\usage{
SoftMax(x)
}
\arguments{
\item{x}{input value (numeric)}
}
\value{
output value (numeric)
}
\description{
SoftMax sigmoid function
}
\examples{
# create a vector of integers (as works for non-integers)
set.seed(0)
V = c(floor(runif(9, min=-3, max=3)))
print(V)

# this outputs
# [1]  2 -2 -1  0  2 -2  2  2  0

# apply the SoftMax() function to V
sV <- SoftMax(V)
print(sV)

# this outputs
# [1] 0.229511038 0.004203641 0.011426682 0.031060941
# 0.229511038 0.004203641 0.229511038 0.229511038 0.031060941

}
