% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeWeights.R
\name{ComputeWeights}
\alias{ComputeWeights}
\title{SoftMax sigmoid function}
\usage{
ComputeWeights(scores)
}
\arguments{
\item{scores}{input value (numeric)}
}
\value{
output value (numeric)
}
\description{
SoftMax sigmoid function
}
